Enable face culling with sodium
